Working on flycheck/flymake
I looked at the two to see if it was worth trying to consolidate on one
over the other.

My impressions are that flycheck does jshint better but that for python
flymake is better integrated into elpy and we have some custom code for
ignore certain messages.

So really just load both.