Midnight: Kill clean `^/cifs/` buffers.
Midnight mode has the `'clean-buffer-list-kill-regexps` variable that
will hurry-kill buffers.

Our normal setting for midnight is to kill buffers that haven't been
used for 5 days. If it matches a regexp in the list
`'clean-buffer-list-kill-regexps` then it will be killed after 1h of
non-use.

This is really useful for our automount that freaks out if the ticket
expires and then we go to access it.

re Programming:#769 (0.25)